Use a MC evaluator class to allow user customization when evaluating energies
The MC class can now take an (optional) evaluate class, in order to allow higher degrees
of customization when running MC simulations, as discussed in issue #268.

Allow either Atoms object or an MC Evaluator object to be passed into the montecarlo constructor.
Use a baseclass for MC and KMC, since they both have the same handling of evaluator and atoms object.

Update docs to new evaluator API, verify as doctest