ENH: Add synchronization method to MC Evaluator
The slow tests were failing, due to a de-synchronization of changes
to the internal atoms object in the CLEASE calculator, and the atoms
in the Evaluator. This adds a "synchronize" method, which should
be called at the beginning of a new MC run, to ensure
things are in sync.

Added test of the new syncronize function for CEMCEvaluator.