Use FourVector objects internally when constructing cluster_list and trans_matrix
Add function in settings to print table of clusters.

Fixes a bug, where the cartesian coordinates would be calculated based on the index in the primitive, which would not match the tag.
The primitive is now sorted according to the order of the tag.

Added tests which test the trans matrix against a previously calculated state, so we know if we make any changes that cause the trans matrix to change values.

Make ClusterManager in charge of masking background atoms.

Only have 1 c0 cluster.

Define the Figure class, which is a container of FourVector objects.
FourVector implements .shift_xyz(other)