BUG: Clusters were not being cut off properly.
The diameter of the clusters were considered in two different ways, and therefore clusters which were too large were considered within the cutoff sphere, which results in some confusion when looking at the cluster table.
Clusters are now only allowed to be used if they also obey being within the geometric diameter cutoff.

Calculate cluster fingerprint with pure NumPy.

Verify the fingerprint in the ClusterFingerprint class